Extracted StopTable into its own component
