+++ src/components/sidebar/nodeView/stopAreaView/stopTable.scss
@@ -1,0 +1,38 @@
+@import '../../../shared/styles/common';
+@import '../../../shared/styles/form';
+
+.stopTableView {
+    width: 100%;
+
+    .stopHeaderTable {
+        width: 100%;
+        font-size: $smallFontSize;
+        border-collapse: separate;
+        border-spacing: 0 5px; /* Padding top/bottom between table rows */
+
+        td {
+            padding-left: 10px;
+        }
+        td:first-of-type {
+            padding: 0px;
+        }
+        .columnHeader {
+            white-space: nowrap;
+            text-align: start;
+            padding-left: 10px;
+        }
+        .columnHeader:first-of-type {
+            padding-left: 0px;
+        }
+    }
+
+    .stopTableRow {
+        font-size: 0.9rem;
+    }
+
+    .sectionHeader {
+        margin-top: 15px;
+        margin-bottom: 15px;
+        font-weight: bold;
+    }
+}
